configurable drawMode property and showing area of drawing
