+++ src/draw.ts
@@ -1,5 +1,7 @@
+import Geometry from "ol/geom/Geometry";
 import { Vector as VectorLayer } from "ol/layer";
 import { Vector as VectorSource } from "ol/source";
+import { getArea } from "ol/sphere";
 import { Circle as CircleStyle, Fill, Stroke, Style } from "ol/style";
 
 const redLineStroke = new Stroke({
@@ -24,3 +26,15 @@ export const drawingLayer = new VectorLayer({
     }),
   }),
 });
+
+// calculate and format the area of a polygon
+export function formatArea(polygon: Geometry) {
+  const area = getArea(polygon);
+  let output;
+  if (area > 10000) {
+    output = Math.round((area / 1000000) * 100) / 100 + " " + "km<sup>2</sup>";
+  } else {
+    output = Math.round(area * 100) / 100 + " " + "m<sup>2</sup>";
+  }
+  return output;
+};
